fix: rename @types → types
