+++ packages/takudev-web/components/providers/ColorProvider/index.tsx
@@ -1,8 +1,8 @@
 import React, { useMemo } from 'react';
 import isEqual from 'react-fast-compare';
 
-import type { ColorMode } from '~/@types';
 import { ColorContext } from '~/contexts';
+import type { ColorMode } from '~/types';
 
 import { useColorProvider } from './useColorProvider';
